Clean up Phase 2 analysis code
No behaviour change to the measurements.

  * excess.py: the paint-error closure was redefined on every loop
    iteration over loop-local variables. That was correct, since it is
    called immediately, but it reads as the classic late-binding bug, so
    it is now a module-level _paint_error. rounds now counts only rounds
    that removed a path.
  * analyze_pairs.py: the per-image and per-setting frontier construction
    was duplicated across three summaries, and is now shared
    _image_fronts / _setting_points. Moved an inline import to the top.
    Dropped messy_rmse_row, which duplicated messy_rmse. The ladder
    length is no longer hardcoded as 5.
  * build_pairs.read_pairs: renamed the filter argument to where, so it
    no longer shadows the builtin.
  * frontier.FrontierPoint.tokens is typed float: aggregate curves hold
    medians.
  * Removed unused imports (download.py, test_frontier.py), made zips
    strict, and dropped unused loop variables in test_tokenizer.py.

+++ im2vec/frontier.py
@@ -58,7 +58,7 @@ SWEEP: List[Tuple[str, Dict[str, int]]] = [
 @dataclass(frozen=True)
 class FrontierPoint:
     label: str
-    tokens: int
+    tokens: float  # an int per trace; a median or mean on aggregate curves
     rmse: float
     paths: int = 0
 
@@ -91,7 +91,7 @@ def pareto_front(points: Sequence[FrontierPoint]) -> List[FrontierPoint]:
     return front
 
 
-def tokens_at(front: Sequence[FrontierPoint], max_rmse: float) -> Optional[int]:
+def tokens_at(front: Sequence[FrontierPoint], max_rmse: float) -> Optional[float]:
     """Fewest tokens on the frontier with RMSE at or below ``max_rmse``."""
     meeting = [p.tokens for p in front if p.rmse <= max_rmse]
     return min(meeting) if meeting else None
